Moved app to its own file so it can be exported for testing
